external-api-app/requirements.txt
fastapi
uvicorn
opencensus-ext-azure